backend/requirements.txt
fastapi==0.115.0
sqlmodel==0.0.22
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.9
uvicorn[standard]==0.30.0
pydantic-settings==2.5.0
python-multipart==0.0.9
